arxiv_podcast/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
openai==1.70.0
packaging==24.2
pluggy==1.5.0
pydantic==2.11.1
pydantic_core==2.33.0
pytest==8.3.5
PyYAML==6.0.2
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
